Test data files moved to tests_data folder and dependencies for test data files updated at test files.
