+++ .gitignore
@@ -1,2 +1,3 @@
 .env
 internal/*
+internal_tests
